+++ src/main/java/edu/ie3/datamodel/io/source/IdCoordinateSource.java
@@ -8,14 +8,11 @@ package edu.ie3.datamodel.io.source;
 import java.util.Collection;
 import org.locationtech.jts.geom.Point;
 
-public interface CoordinateSource extends DataSource {
+public interface IdCoordinateSource extends DataSource {
 
   Point getCoordinate(Integer id);
 
   Collection<Point> getCoordinates(Integer... ids);
-
-  // TODO Remove after Projektarbeit
-  Collection<Point> getCoordinatesBetween(Integer fromId, Integer toId);
 
   Integer getId(Point coordinate);
 }