+++ src/main/java/edu/ie3/datamodel/io/source/influxdb/InfluxDbWeatherSource.java
@@ -8,7 +8,7 @@ package edu.ie3.datamodel.io.source.influxdb;
 import edu.ie3.datamodel.io.connectors.InfluxDbConnector;
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedWeatherValueData;
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedWeatherValueFactory;
-import edu.ie3.datamodel.io.source.CoordinateSource;
+import edu.ie3.datamodel.io.source.IdCoordinateSource;
 import edu.ie3.datamodel.io.source.WeatherSource;
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue;
@@ -23,18 +23,30 @@ import org.influxdb.dto.Query;
 import org.influxdb.dto.QueryResult;
 import org.locationtech.jts.geom.Point;
 
+/** InfluxDB Source for weather data */
 public class InfluxDbWeatherSource implements WeatherSource {
-  private static final String COORDINATE_ID = "coordinate";
+  private static final String BASIC_QUERY_STRING = "Select * from weather";
+  private static final String COORDINATE_ID_COLUMN_NAME = "coordinate";
+  private static final String MEASUREMENT_NAME_WEATHER = "weather";
+  private static final int MILLI_TO_NANO_FACTOR = 1000000;
   private final InfluxDbConnector connector;
-  private final CoordinateSource coordinateSource;
+  private final IdCoordinateSource coordinateSource;
   private final TimeBasedWeatherValueFactory weatherValueFactory;
 
-  public InfluxDbWeatherSource(InfluxDbConnector connector, CoordinateSource coordinateSource) {
+  /**
+   * Initalizes a new InfluxDbWeatherSource
+   *
+   * @param connector needed for database connection
+   * @param coordinateSource needed to map coordinates to ID as InfluxDB does not support spatial
+   *     types
+   */
+  public InfluxDbWeatherSource(InfluxDbConnector connector, IdCoordinateSource coordinateSource) {
     this.connector = connector;
     this.coordinateSource = coordinateSource;
     this.weatherValueFactory = new TimeBasedWeatherValueFactory();
   }
 
+  @Override
   public Map<Point, IndividualTimeSeries<WeatherValue>> getWeather(
       ClosedInterval<ZonedDateTime> timeInterval) {
     try (InfluxDB session = connector.getSession()) {
@@ -76,7 +88,14 @@ public class InfluxDbWeatherSource implements WeatherSource {
     return coordinateToTimeSeries;
   }
 
-  public IndividualTimeSeries<WeatherValue> getWeatherForCoordinate(
+  /**
+   * Return the weather for the given time interval AND coordinate
+   *
+   * @param timeInterval Queried time interval
+   * @param coordinate Queried coordinate
+   * @return weather data for the specified time and coordinate
+   */
+  public IndividualTimeSeries<WeatherValue> getWeather(
       ClosedInterval<ZonedDateTime> timeInterval, Point coordinate) {
     try (InfluxDB session = connector.getSession()) {
       String query = createQueryStringForIntervalAndCoordinate(timeInterval, coordinate);
@@ -96,53 +115,54 @@ public class InfluxDbWeatherSource implements WeatherSource {
     }
   }
 
+  /**
+   * Parses an influxQL QueryResult and then transforms them into a Stream of optional
+   * TimeBasedValue&lt;WeatherValue&gt;, with a present Optional value, if the transformation was
+   * successful and an empty optional otherwise.
+   */
   private Stream<Optional<TimeBasedValue>> optTimeBasedValueStream(QueryResult queryResult) {
     Map<String, Set<Map<String, String>>> measurementsMap =
-        InfluxDbConnector.parseQueryResult(queryResult, "weather");
-    return measurementsMap.get("weather").stream()
+        InfluxDbConnector.parseQueryResult(queryResult, MEASUREMENT_NAME_WEATHER);
+    return measurementsMap.get(MEASUREMENT_NAME_WEATHER).stream()
         .map(
             fields ->
                 new TimeBasedWeatherValueData(
                     fields,
-                    coordinateSource.getCoordinate(Integer.valueOf(fields.remove(COORDINATE_ID)))))
+                    coordinateSource.getCoordinate(
+                        Integer.valueOf(fields.remove(COORDINATE_ID_COLUMN_NAME)))))
         .map(weatherValueFactory::getEntity);
   }
 
-  public String createQueryStringForIntervalAndCoordinate(
+  private String createQueryStringForIntervalAndCoordinate(
       ClosedInterval<ZonedDateTime> timeInterval, Point coordinate) {
     return createQueryStringForInterval(timeInterval)
         + " and "
         + createCoordinateConstraintString(coordinate);
   }
 
-  public String createQueryStringForDateAndCoordinate(ZonedDateTime date, Point coordinate) {
+  private String createQueryStringForDateAndCoordinate(ZonedDateTime date, Point coordinate) {
     return createQueryStringForDate(date) + " and " + createCoordinateConstraintString(coordinate);
   }
 
-  public String createQueryStringForInterval(ClosedInterval<ZonedDateTime> timeInterval) {
-    String basicQuery = createBasicQueryString();
+  private String createQueryStringForInterval(ClosedInterval<ZonedDateTime> timeInterval) {
     String timeConstraint =
         "time >= "
-            + timeInterval.getLower().toInstant().toEpochMilli() * 1000000
+            + timeInterval.getLower().toInstant().toEpochMilli() * MILLI_TO_NANO_FACTOR
             + " and time <= "
-            + timeInterval.getUpper().toInstant().toEpochMilli() * 1000000;
-    return basicQuery + " where " + timeConstraint;
+            + timeInterval.getUpper().toInstant().toEpochMilli() * MILLI_TO_NANO_FACTOR;
+    return BASIC_QUERY_STRING + " where " + timeConstraint;
   }
 
-  public String createQueryStringForDate(ZonedDateTime date) {
-    String basicQuery = createBasicQueryString();
-    String timeConstraint = "time=" + date.toInstant().toEpochMilli() * 1000000;
-    return basicQuery + " where " + timeConstraint;
+  private String createQueryStringForDate(ZonedDateTime date) {
+    String timeConstraint = "time=" + date.toInstant().toEpochMilli() * MILLI_TO_NANO_FACTOR;
+    return BASIC_QUERY_STRING + " where " + timeConstraint;
   }
 
-  public String createBasicQueryString() {
-    return "Select * from weather";
-  }
-
-  public String createCoordinateConstraintString(Point coordinate) {
+  private String createCoordinateConstraintString(Point coordinate) {
     return "coordinate='" + coordinateSource.getId(coordinate) + "'";
   }
 
+  /** Removes empty Optionals */
   protected Stream<TimeBasedValue<WeatherValue>> filterEmptyOptionals(
       Stream<Optional<TimeBasedValue>> elements) {
     return elements.filter(Optional::isPresent).map(Optional::get).map(TimeBasedValue.class::cast);