+++ src/main/java/edu/ie3/datamodel/models/result/system/HpResult.java
@@ -1,0 +1,76 @@
+/*
+ * © 2020. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+*/
+package edu.ie3.datamodel.models.result.system;
+
+import java.time.ZonedDateTime;
+import java.util.Objects;
+import java.util.UUID;
+import javax.measure.Quantity;
+import javax.measure.quantity.Power;
+import tec.uom.se.ComparableQuantity;
+
+/** Represents calculation results of a {@link edu.ie3.datamodel.models.input.system.HpInput} */
+public class HpResult extends SystemParticipantResult {
+
+  private final Quantity<Power> qDot;
+
+  /**
+   * Standard constructor with automatic uuid generation.
+   *
+   * @param timestamp date and time when the result is produced
+   * @param inputModel uuid of the input model that produces the result
+   * @param p active power output normally provided in MW
+   * @param q reactive power output normally provided in MVAr
+   */
+  public HpResult(
+      ZonedDateTime timestamp,
+      UUID inputModel,
+      ComparableQuantity<Power> p,
+      ComparableQuantity<Power> q,
+      Quantity<Power> qDot) {
+    super(timestamp, inputModel, p, q);
+    this.qDot = qDot;
+  }
+
+  /**
+   * Standard constructor which allows uuid provision
+   *
+   * @param uuid uuid of this result entity, for automatic uuid generation use primary constructor
+   *     above
+   * @param timestamp date and time when the result is produced
+   * @param inputModel uuid of the input model that produces the result
+   * @param p active power output normally provided in MW
+   * @param q reactive power output normally provided in MVAr
+   */
+  public HpResult(
+      UUID uuid,
+      ZonedDateTime timestamp,
+      UUID inputModel,
+      ComparableQuantity<Power> p,
+      ComparableQuantity<Power> q,
+      Quantity<Power> qDot) {
+    super(uuid, timestamp, inputModel, p, q);
+    this.qDot = qDot;
+  }
+
+  public Quantity<Power> getqDot() {
+    return qDot;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    if (!super.equals(o)) return false;
+    HpResult hpResult = (HpResult) o;
+    return qDot.equals(hpResult.qDot);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(super.hashCode(), qDot);
+  }
+}