+++ src/test/groovy/edu/ie3/datamodel/io/source/influxdb/InfluxDbWeatherSourceIT.groovy
@@ -20,7 +20,7 @@ import spock.lang.Shared
 import spock.lang.Specification
 
 @Testcontainers
-class InfluxDbWeatherSourceTest extends Specification implements WeatherSourceTestHelper {
+class InfluxDbWeatherSourceIT extends Specification implements WeatherSourceTestHelper {
 
 	@Shared
 	InfluxDBContainer influxDbContainer = new InfluxDBContainer()
@@ -32,6 +32,8 @@ class InfluxDbWeatherSourceTest extends Specification implements WeatherSourceTe
 	InfluxDbWeatherSource source
 
 	def setupSpec() {
+		// Copy import file into docker and then import it via influx CLI
+		// more information on file format and usage here: https://docs.influxdata.com/influxdb/v1.7/tools/shell/#import-data-from-a-file-with-import
 		MountableFile influxWeatherImportFile = MountableFile.forClasspathResource("/testcontainersFiles/influxDb/weather.txt");
 		influxDbContainer.copyFileToContainer(influxWeatherImportFile, "/home/weather.txt")
 		def execResult = influxDbContainer.execInContainer("influx", "-import", "-path=/home/weather.txt", "-precision=ms")